+++ packages/review/src/components/infinity-list/index.ts
@@ -1,1 +1,2 @@
 export * from './popular-reviews-infinite'
+export * from './latest-reviews-infinite'
